Sincerus_Website
Lots of test work included in this folder. Instead of using a .png I used CSS to style particular aspects of the pages. Website is not finished but is running. Unfortunately client has not filled in some of the forms required to finish the website. It's my first true website.